+++ frontend/src/App.jsx
@@ -9,10 +9,10 @@ import { createBrowserRouter, Outlet, RouterProvider } from 'react-router-dom';
 import Navigation from './components/Navigation/Navigation';
 import MobileNotice from './components/MobileNotice/MobileNotice';
 import LandingPage from './components/Landing/Landing';
-import RecentLoadouts from './components/Loadouts/RecentLoadouts';
+import RecentLoadouts from './components/LoadoutLists/RecentLoadouts';
 import LoadoutBuilderMain from './components/LoadoutBuilder/LoadoutBuilder';
 import UserProfile from './components/Profile/Profile';
-import UserLoadouts from './components/Loadouts/UserLoadouts';
+import UserLoadouts from './components/LoadoutLists/UserLoadouts';
 import Footer from './components/Footer/Footer';
 import useWindowSize from './hooks/useWindowSize';
 import * as sessionActions from './store/session';
