+++ frontend/src/components/LoadoutLists/LoadoutList.css
@@ -1,0 +1,134 @@
+/* * frontend/src/components/Loadouts/LoadoutList.css */
+/* * Site Loadout List */
+
+/* Site Loadout List - Global Definitions */
+.site-loadout-list {
+    backdrop-filter: brightness(50%);
+    border: 2px solid #acdbf9;
+    box-sizing: content-box;
+    display: grid;
+    margin: auto;
+    min-height: max-content;
+    padding: 0 10px;
+    width: 100%;
+}
+
+/* Loadout List: Single Loadout */
+.site-single-loadout {
+    backdrop-filter: brightness(50%);
+    border: 2px solid #acdbf9;
+    box-sizing: border-box;
+    display: grid;
+        column-gap: 1vw;
+        grid-template-columns: 22.5% auto fit-content(22.5%);
+    margin: 10px 0;
+    height: fit-content;
+    padding: 1vw 1.5vw;
+    z-index: 1;
+}
+.site-single-loadout:hover {
+    border: 2px solid #f9e2ad;
+    cursor: pointer;
+}
+
+/* Single Loadout: Loadout Info (Title, Ship Name) */
+.single-loadout-info {
+    display: flex;
+        flex-direction: column;
+        row-gap: 0.5vw;
+    height: 100%;
+    overflow: hidden;
+    pointer-events: none;
+}
+.single-loadout-info h2 { 
+    font-size: 1.85vw;
+    margin: 0; 
+}
+.single-loadout-info__ship {
+    display: inline-flex;
+        align-items: center;
+    font-weight: 600;
+    font-size: 1.15vw;
+    letter-spacing: 0.75px;
+    margin-top: 5px;
+    text-transform: uppercase;
+}
+.single-loadout-info__ship img { 
+    height: 2vw; 
+    margin-right: 5px;
+}
+.single-loadout-info__enhancements {
+    border: 2px solid rgba(148, 180, 195, 0.7);
+        border-radius: 5px;
+    display: flex;
+        column-gap: 1vw;
+        justify-content: center;
+    height: fit-content;
+        min-height: 2.5vw;
+    margin: 10px 0 7px 0;
+    padding: 0.5vw 0;
+}
+.single-loadout-info__enhancements img {
+    aspect-ratio: 1 / 1;
+    backdrop-filter: brightness(40%);
+    border: 2px solid rgba(148, 180, 195, 0.7);
+        border-radius: 5px;
+    height: 2.5vw;
+    padding: 0.25vw;
+}
+
+.single-loadout-list {
+    height: 100%;
+    overflow: hidden;
+    pointer-events: none;
+}
+.single-loadout-list ul {
+    display: grid;
+    list-style: none;
+    margin: 0;
+    padding-left: 0;
+    row-gap: 10px;
+}
+.single-loadout-list li {
+    display: flex;
+        column-gap: 10px;
+        justify-content: center;
+    height: 4vw;
+}
+.single-loadout-list img {
+    aspect-ratio: 13 / 9;
+    backdrop-filter: brightness(40%);
+    border: 2px solid rgba(148, 180, 195, 0.7);
+}
+
+.single-loadout-owner {
+    display: inline-flex;
+        align-items: flex-end;
+        flex-direction: column;
+    height: 100%;
+}
+.single-loadout-owner a {
+    font-size: 1.2vw;
+    font-weight: 700;
+    letter-spacing: 0.75px;
+    position: relative;
+    text-align: right;
+    text-transform: uppercase;
+    z-index: 2;
+}
+.single-loadout-owner button {
+    font-size: 1.8vw;
+    position: relative;
+    z-index: 2;
+}
+.single-loadout-owner p {
+    font-size: 1vw;
+    letter-spacing: 0.5px;
+    margin: 0;
+    pointer-events: none;
+    text-align: right;
+}
+
+.single-loadout-owner__ctrls {
+    width: fit-content;
+}
